+++ core/src/main/java/react4j/core/NativeAdapterComponent.java
@@ -60,7 +60,7 @@ public abstract class NativeAdapterComponent<
     }
     try
     {
-      _component.componentDidConstruct();
+      _component.performComponentDidConstruct();
     }
     finally
     {