+++ processor/src/main/java/react4j/processor/Constants.java
@@ -8,6 +8,7 @@ final class Constants
   static final String EVENT_HANDLER_ANNOTATION_CLASSNAME = "react4j.annotations.EventHandler";
   static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";
   static final String INJECT_ANNOTATION_CLASSNAME = "javax.inject.Inject";
+  static final String AUTORUN_ANNOTATION_CLASSNAME = "org.realityforge.arez.annotations.Autorun";
 
   private Constants()
   {